requirements.txt
accelerate==0.29.3
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.5.0
chromedriver-autoinstaller==0.6.4
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
cssselect==1.2.0
dataclasses-json==0.6.5
datasets==2.19.0
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
einops==0.8.0
etelemetry==0.3.1
executing==2.0.1
fastapi==0.110.3
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
fitz==0.0.1.dev2
flatbuffers==24.3.25
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kubernetes==29.0.0
langchain==0.1.17
langchain-chroma==0.1.1
langchain-community==0.0.36
langchain-core==0.1.48
langchain-text-splitters==0.0.1
langsmith==0.1.52
llama-index==0.10.33
llama-index-agent-openai==0.2.3
llama-index-cli==0.1.12
llama-index-core==0.10.34
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.3
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-index-readers-web==0.1.12
llama-index-vector-stores-chroma==0.1.7
llama-parse==0.4.2
llamaindex-py-client==0.1.19
llmsherpa==0.1.4
looseversion==1.3.0
loralib==0.1.2
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==2.0.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.25.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
playwright==1.43.0
posthog==3.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
prov==2.0.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydot==2.0.0
pyee==11.1.0
Pygments==2.17.2
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==26.0.3
rdflib==7.0.0
regex==2024.4.28
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.20.0
sentence-transformers==2.7.0
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
traits==6.3.2
transformers @ git+https://github.com/huggingface/transformers@4ad5adaf1d224fa28ffa8e1d124846b1d55a5d0e
trio==0.25.0
trio-websocket==0.11.1
triton==2.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
